requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.3
django-cors-headers==4.3.1
django-ninja==1.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.0
iniconfig==2.0.0
kombu==5.3.6
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
